+++ lib/angola_province_picker.dart
@@ -1,34 +1,15 @@
-/// Um seletor de províncias de Angola personalizável para Flutter.
+/// Biblioteca principal para o Angola Province Picker.
 ///
-/// O `AngolaProvincePicker` é um widget reutilizável que fornece uma interface
-/// simples e amigável para selecionar uma das 18 províncias de Angola.
-///
-/// Ideal para formulários de cadastro, endereços e aplicativos voltados ao público angolano.
-///
-/// Recursos principais:
-/// - Lista completa das províncias angolanas.
-/// - Suporte a modos Material e Cupertino.
-/// - Totalmente personalizável.
-/// - Fácil de integrar com `Form` e `TextFormField`.
-///
-/// Exemplo de uso:
-/// ```dart
-/// AngolaProvincePicker(
-///       selected: selected,
-///    onChanged: (value) {
-///       setState(() {
-///         selected = value;
-///        print(value);
-///        });
-///      },
-///      hint: 'Escolha uma provincia',
-///      dropdownColor: Colors.white,
-///    ),
-/// ```
-///
-/// Desenvolvido para facilitar a vida de desenvolvedores que constroem apps localizados.
-
+/// Este pacote fornece widgets reutilizáveis para selecionar
+/// províncias e municípios de Angola em estilo Material ou Cupertino.
 library;
 
-export 'src/provinces.dart';
-export 'src/widgets/material_province_picker.dart';
+export 'src/models/towns.dart';
+export 'src/models/provinces.dart';
+export 'src/widgets/angola_province_picker.dart';
+export 'src/utils/angola_picker_mode.dart';
+export 'src/widgets/cupertino/cupertino_province_picker.dart';
+export 'src/widgets/material/material_province_picker.dart';
+export 'src/widgets/angola_town_picker.dart';
+export 'src/widgets/cupertino/cupertino_town_picker.dart';
+export 'src/widgets/material/material_town_picker.dart';